UCT-704: Refactored deprecated inspections, added base class for target inspection
