+++ src/com/magento/idea/magento2uct/inspections/php/deprecation/UsingDeprecatedInterface.java
@@ -7,47 +7,30 @@ package com.magento.idea.magento2uct.inspections.php.deprecation;
 
 import com.intellij.codeInspection.ProblemHighlightType;
 import com.intellij.codeInspection.ProblemsHolder;
+import com.intellij.openapi.project.Project;
 import com.jetbrains.php.lang.psi.elements.ClassReference;
-import com.jetbrains.php.lang.psi.elements.Field;
+import com.jetbrains.php.lang.psi.elements.PhpClass;
 import com.magento.idea.magento2uct.inspections.UctProblemsHolder;
+import com.magento.idea.magento2uct.inspections.php.UsedTypeInspection;
 import com.magento.idea.magento2uct.packages.IssueSeverityLevel;
 import com.magento.idea.magento2uct.packages.SupportedIssue;
+import com.magento.idea.magento2uct.versioning.VersionStateManager;
 import org.jetbrains.annotations.NotNull;
 
-public class UsingDeprecatedInterface extends UsingDeprecatedType {
+public class UsingDeprecatedInterface extends UsedTypeInspection {
 
     @Override
-    protected void registerProblem(
+    protected void execute(
+            final Project project,
             final @NotNull ProblemsHolder problemsHolder,
-            final Field field,
-            final String fieldInterface,
-            final boolean isInterface
+            final PhpClass phpClass,
+            final ClassReference reference
     ) {
-        if (!isInterface) {
+        if (!phpClass.isInterface()
+                || !VersionStateManager.getInstance(project).isDeprecated(phpClass.getFQN())) {
             return;
         }
-        if (problemsHolder instanceof UctProblemsHolder) {
-            ((UctProblemsHolder) problemsHolder).setReservedErrorCode(
-                    SupportedIssue.USING_DEPRECATED_INTERFACE.getCode()
-            );
-        }
-        problemsHolder.registerProblem(
-                field,
-                SupportedIssue.USING_DEPRECATED_INTERFACE.getMessage(fieldInterface),
-                ProblemHighlightType.LIKE_DEPRECATED
-        );
-    }
 
-    @Override
-    protected void registerReferenceProblem(
-            final @NotNull ProblemsHolder problemsHolder,
-            final ClassReference reference,
-            final String deprecatedType,
-            final boolean isInterface
-    ) {
-        if (!isInterface) {
-            return;
-        }
         if (problemsHolder instanceof UctProblemsHolder) {
             ((UctProblemsHolder) problemsHolder).setReservedErrorCode(
                     SupportedIssue.USING_DEPRECATED_INTERFACE.getCode()
@@ -55,7 +38,7 @@ public class UsingDeprecatedInterface extends UsingDeprecatedType {
         }
         problemsHolder.registerProblem(
                 reference,
-                SupportedIssue.USING_DEPRECATED_INTERFACE.getMessage(deprecatedType),
+                SupportedIssue.USING_DEPRECATED_INTERFACE.getMessage(phpClass.getFQN()),
                 ProblemHighlightType.LIKE_DEPRECATED
         );
     }
